# api-generator/api-definition.yaml
openapi: 3.0.3
info:
  title: WadLoader3Api
  description: WadLoader3Api
  version: 1.0.0
servers:
  - url: "http://localhost:8080"
...
paths:
  /wad:
    post:
      operationId: postWad
      description: "Create a new Wad"
      tags:
        - wad
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/NewWadDto"
      responses:
        200:
          description: "Uploaded new Wad"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WadDto'
        400:
          description: "No Wad uploaded"
...
components:
  schemas:
    WadDto:
      type: object
      required:
        - id
        - name
        - description
      properties:
        name:
          type: string
        description:
          type: string
        id:
          type: integer
...
    NewWadDto:
      type: object
      required:
        - name
        - description
        - file
      properties:
        name:
          type: string
        description:
          type: string
        file:
          type: string
          format: binary
...